UI responsiveness (#249)
* Add invalid input responses to Brute Force

* Add invalid input responses to Boyer Moore

* Add invalid input responses to KMP

* Add invalid input responses to Rabin Karp

* Fix text alignment issue in StackArray popping

Popped element now displays in same location as pushed elements rather than being concatenated to the label string, which causes alignment issues

* Fix text alignment issue in popping for StackLL, QueueArray, QueueLL

* Fix various inconsistencies in Dequeues

* Add invalid input responses to BuildHeap

* Fix Load Factor text alignment & input

Even though the text box allowed for 5 digits of precision for load factor, only a 1 or 2-digit number was considered valid in the code. I reduced the text box to 3 digits (I see no reason for more) and fixed the validation code.

* Prettier formatting

* Make LF inputs/display consistent in both HashMaps

* Add invalid input responses to Graph algos

* Add invalid input responses to sorts

* Add invalid input responses & length display to LCS

* Fix info label clear on HeapSort start

* Clear button resets text input fields

* Make text input fields persistent when algo is run

This change makes it so input fields for pattern matching and sorts will retain the data typed into them, allowing for easier re-runs and editing. The fields will be cleared when the clear button is pressed.

* Add Clear button to LCS

* Prettier formatting

* Fix resetAll calls in HashMaps

* Add Length of LCS label & fix redundancy

* Fix nextIndex increment in Deque/Queue/Stack Arrays